requirements.txt
streamlit
pandas
matplotlib
seaborn
requests
python-dotenv
groq
numpy